requirements.txt
jinja2==3.0.1
google==3.0.0
google-api-core==2.8.0
google-auth==2.6.6
google-cloud-pubsub==2.12.1
googleapis-common-protos==1.56.1
grpc-google-iam-v1==0.12.4
mysql
mysqlclient==1.4.6
mysql-connector-python==8.0.23
numpy==1.19.0
pandas==1.3.5
protobuf==3.20.1
pydantic==1.9.0
pyjwt==2.4.0
pymongo[srv]==4.1.1
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.3.16
typing_extensions==3.7.4.3
uk_postcode_utils==1.0
validators==0.18.2